VOA-2561 - Tidy up code and update contact details to full name
