+++ app/uk/gov/hmrc/lightweightcontactevents/models/ConfirmedContactDetails.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 HM Revenue & Customs
+ * Copyright 2021 HM Revenue & Customs
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,8 +18,7 @@ package uk.gov.hmrc.lightweightcontactevents.models
 
 import play.api.libs.json.Json
 
-case class ConfirmedContactDetails(firstName: String,
-                                   lastName: String,
+case class ConfirmedContactDetails(fullName: String,
                                    email: String,
                                    contactNumber: String
                                   )
