+++ conf/application.conf
@@ -1,4 +1,4 @@
-# Copyright 2020 HM Revenue & Customs
+# Copyright 2021 HM Revenue & Customs
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
